requirements.txt
PyYAML~=6.0.2
pillow~=11.1.0
ultralytics~=8.3.67